refactor(prize): 기존 Prize 클래스를 Rank클래스로 수정 및 관련 테스트 코드 삭제
