+++ src/main/java/lotto/domain/Rank.java
@@ -1,0 +1,35 @@
+package lotto.domain;
+
+import java.util.Arrays;
+
+public enum Rank {
+    FIRST(6, 2_000_000_000),
+    SECOND(5, 1_500_000),
+    THIRD(4, 50_000),
+    FOURTH(3, 5_000);
+
+    private final int matchCount;
+
+    private final int prize;
+
+    Rank(int matchCount, int prize) {
+        this.matchCount = matchCount;
+        this.prize = prize;
+    }
+
+    public int getMatchCount() {
+        return matchCount;
+    }
+
+    public int getPrize() {
+        return prize;
+    }
+
+    public static int getRankingPrize(int matchCount) {
+        return Arrays.stream(Rank.values())
+                .filter(rank -> rank.matchCount == matchCount)
+                .findAny()
+                .map(Rank::getPrize)
+                .orElse(0);
+    }
+}
